NvidiaTracker/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.0",
    "beautifulsoup4>=4.13.3",
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "lxml>=5.3.2",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.3",
    "scheduler>=0.8.8",
    "trafilatura>=2.0.0",
]
